Added css to sepa mandate page and removed cancel button
