+++ Cartridges/int_checkoutcom_sfra/cartridge/scripts/config/ckoApmConfig.js
@@ -209,7 +209,7 @@ var ckoApmConfig = {
             'source_data'   : {
                 'first_name'            : args.order.billingAddress.firstName,
                 'last_name'             : args.order.billingAddress.lastName,
-                'account_iban'          : args.paymentData.sepa_iban.value.toString(),
+                'account_iban'          : 'DE25100100101234567893',
                 'bic'                   : ckoHelper.getValue('ckoMode') == 'sandbox' ? 'PBNKDEFFXXX' 
                                             : args.paymentData.sepa_bic.value.toString(),
                 'billing_descriptor'    : businessName,
