+++ Cartridges/int_checkoutcom_sfra/cartridge/scripts/helpers/cardHelper.js
@@ -2,12 +2,10 @@
 
 /* API Includes */
 var OrderMgr = require('dw/order/OrderMgr');
-var Transaction = require('dw/system/Transaction');
-var CustomerMgr = require('dw/customer/CustomerMgr');
-var URLUtils = require('dw/web/URLUtils');
 
 /** Utility **/
 var ckoHelper = require('~/cartridge/scripts/helpers/ckoHelper');
+var savedCardHelper = require('~/cartridge/scripts/helpers/savedCardHelper');
 
 /*
 * Utility functions for my cartridge integration.
@@ -92,7 +90,7 @@ var cardHelper = {
         // Handle the save card request
         if (paymentData.creditCardFields.saveCard.value) {
             // Save the card
-            var uuid = this.saveCard(
+            var uuid = savedCardHelper.saveCard(
                 paymentData,
                 req
             );
@@ -120,7 +118,7 @@ var cardHelper = {
         var condition2 = selectedCardUuid && selectedCardUuid.length > 0;
         if (condition1 && condition2) {
             // Get the saved card
-            var savedCard = this.getSavedCard(
+            var savedCard = savedCardHelper.getSavedCard(
                 selectedCardUuid.toString(),
                 order.getCustomerNo(),
                 processorId
@@ -153,163 +151,6 @@ var cardHelper = {
             'enabled' : ckoHelper.getValue('cko3ds'),
             'attempt_n3d' : ckoHelper.getValue('ckoN3ds')
         }
-    },
-
-    /*
-     * Get a customer saved card
-     */
-    getSavedCard: function (cardUuid, customerNo, methodId) {
-        // Get the customer
-        var customer = CustomerMgr.getCustomerByCustomerNumber(customerNo);
-
-        // Get the customer wallet
-        var wallet = customer.getProfile().getWallet();
-
-        // Get the existing payment instruments
-        var paymentInstruments = wallet.getPaymentInstruments(methodId);
-
-        // Match the saved card
-        for (var i = 0; i < paymentInstruments.length; i++) {
-            var card = paymentInstruments[i];
-            if (card.getUUID() == cardUuid) {
-                return card;
-            }
-        } 
-        
-        return null;
-    },
-
-    /*
-     * Get all customer saved cards
-     */
-    getSavedCards: function (customerNo, methodId) {
-        // Prepare the processor id
-        var processorId = methodId || 'CHECKOUTCOM_CARD';
-
-        // Get the customer
-        var customer = CustomerMgr.getCustomerByCustomerNumber(customerNo);
-
-        // Get the customer wallet
-        var wallet = customer.getProfile().getWallet();
-
-        // Get the existing payment instruments
-        var paymentInstruments = wallet.getPaymentInstruments(processorId);
-        
-        // Prepare the return value
-        var cards = [];
-
-        // Match the saved cards
-        for (var i = 0; i < paymentInstruments.length; i++) {
-            var paymentInstrument = paymentInstruments[i];
-            var condition = (processorId) ? paymentInstrument.paymentMethod == processorId : true;
-            if (condition) {
-                // Card data
-                var card = {
-                    creditCardHolder: paymentInstrument.creditCardHolder,
-                    maskedCreditCardNumber: paymentInstrument.maskedCreditCardNumber,
-                    creditCardType: paymentInstrument.creditCardType,
-                    creditCardExpirationMonth: paymentInstrument.creditCardExpirationMonth,
-                    creditCardExpirationYear: paymentInstrument.creditCardExpirationYear,
-                    UUID: paymentInstrument.UUID,
-                    paymentMethod: paymentInstrument.paymentMethod
-                };
-
-                // Card image
-                card.cardTypeImage = {
-                    src: URLUtils.staticURL('/images/' +
-                    paymentInstrument.creditCardType.toLowerCase().replace(/\s/g, '') +
-                        '-dark.svg'),
-                    alt: paymentInstrument.creditCardType
-                };
-
-                cards.push(card);
-            }
-        } 
-
-        return cards;
-    },
-
-    /*
-     * Save a card in customer account
-     */
-    saveCard: function (paymentData, req) {
-        // Get the customer profile
-        var customerNo = req.currentCustomer.profile.customerNo;
-        var customerProfile = CustomerMgr.getCustomerByCustomerNumber(customerNo).getProfile();
-        var processorId = paymentData.paymentMethod.value;
-    
-        // Build the customer full name
-        var fullName = this.getCustomerFullName(customerProfile); 
-    
-        // Get the customer wallet
-        var wallet = customerProfile.getWallet();
-
-        // The return value
-        var uuid;
-        
-        // Create a stored payment instrument
-        Transaction.wrap(function () {
-            var storedPaymentInstrument = wallet.createPaymentInstrument(processorId);
-            storedPaymentInstrument.setCreditCardHolder(fullName);
-            storedPaymentInstrument.setCreditCardNumber(paymentData.creditCardFields.cardNumber.value);
-            storedPaymentInstrument.setCreditCardExpirationMonth(paymentData.creditCardFields.expirationMonth.value);
-            storedPaymentInstrument.setCreditCardExpirationYear(paymentData.creditCardFields.expirationYear.value);
-            storedPaymentInstrument.setCreditCardType(paymentData.creditCardFields.cardType.value.toLowerCase());
-            uuid = storedPaymentInstrument.getUUID();
-        });
-
-        return uuid;
-    },
-
-    /*
-     * Update a card in customer account
-     */
-    updateSavedCard: function (hook) {
-        var condition1 = hook.data.metadata.hasOwnProperty('card_uuid');
-        var condition2 = hook.data.metadata.hasOwnProperty('customer_id');
-        if (condition1 && condition2) {                
-            // Get the card
-            var card = this.getSavedCard(
-                hook.data.metadata.card_uuid,
-                hook.data.metadata.customer_id,
-                hook.data.metadata.payment_processor
-            );
-        
-            // Create a stored payment instrument
-            if (card) {     
-                Transaction.wrap(function () {
-                    card.setCreditCardToken(hook.data.source.id);
-                });
-            }
-        }
-    },
-
-    getRenderedPaymentInstruments: function (req, accountModel) {
-        var result;
-    
-        if (req.currentCustomer.raw.authenticated
-            && req.currentCustomer.raw.registered
-            && req.currentCustomer.raw.profile.wallet.paymentInstruments.getLength()
-        ) {
-            var context;
-            var template = 'checkout/billing/paymentOptions/forms/savedCardForm';
-    
-            context = { customer: accountModel };
-            result = renderTemplateHelper.getRenderedHtml(
-                context,
-                template
-            );
-        }
-    
-        return result || null;
-    },
-
-    getCustomerFullName: function(customerProfile) { 
-        var customerName = '';
-        customerName += customerProfile.firstName;
-        customerName += ' ' + customerProfile.lastName;
-        
-        return customerName;
     }
 }
