+++ Cartridges/int_checkoutcom_sfra/cartridge/static/default/css/sepa.css
@@ -1,0 +1,91 @@
+
+#cko-sepa .cko-form-container{
+	min-width: 600px; 
+	max-width: 1000px; 
+	border: solid 1px #eee; 
+	border-radius: 5px; 
+	margin: auto; 
+	margin-top: 100px;
+}
+
+#cko-sepa .cko-form-container .cko-top{
+	padding: 40px; 
+	text-align: left; 
+	min-height: 250px;
+}
+
+#cko-sepa .cko-form-container .cko-top .cko-top-title{
+	font-size: 1.5em;
+}
+
+#cko-sepa .cko-form-container .cko-top .cko-sepa-creditor{
+	width: 50%; 
+	float: left;
+}
+
+#cko-sepa .cko-form-container .cko-top .cko-sepa-debtor{
+	width: 50%; 
+	float: right;
+}
+
+#cko-sepa .cko-form-container .cko-top .cko-sepa-creditor h2,
+#cko-sepa .cko-form-container .cko-top .cko-sepa-debtor h2{
+	font-size: 1.3em;
+}
+
+
+#cko-sepa .cko-form-container .cko-top .cko-sepa-creditor p span,
+#cko-sepa .cko-form-container .cko-top .cko-sepa-debtor p span{
+	font-size: 1.3em;
+}
+
+
+#cko-sepa .cko-form-container .cko-bottom{ 
+	text-align: left;
+}
+
+#cko-sepa .cko-form-container .cko-bottom .cko-sepa-message{
+	padding: 20px;
+}
+
+
+#cko-sepa .cko-form-container .cko-bottom .cko-sepa-tandc{
+	padding: 20px;
+}
+
+#cko-sepa .cko-form-container .cko-bottom .cko-sepa-tandc div{
+	//border: solid 1px red;
+	width: 30px;
+	float: left;
+}
+
+
+#cko-sepa .cko-form-container .cko-bottom .cko-sepa-right{
+	padding: 10px; 
+	text-align: center;
+	opacity: 0.3;
+}
+
+#cko-sepa .cko-sepa-button-container{
+	text-align: center; 
+	width: 70%; 
+	max-width: 1100px; 
+	min-width: 600px; 
+	margin: auto; 
+	margin-top: 50px;
+}
+
+#cko-sepa .cko-sepa-button-container .cko-button-group{
+	border: none;
+}
+
+#cko-sepa .cko-sepa-button-container .cko-button-group .cko-sepa-submit-btn{
+	width: 30%; 
+	height: 35px; 
+	background: #ccc;
+}
+
+#cko-sepa .cko-sepa-button-container .cko-button-group .cko-sepa-cancel-btn{
+	width: 20%; 
+	height: 35px;
+}
