+++ Cartridges/int_checkoutcom_sfra/cartridge/templates/default/checkout/billing/paymentOptions/forms/savedCardForm.isml
@@ -1,45 +1,59 @@
 <!--- selected card --->
+<isscript>
+    var BasketMgr = require('dw/order/BasketMgr');
+    var savedCardHelper = require('~/cartridge/scripts/helpers/savedCardHelper');
+</isscript>
 <isloop items="${pdict.customer.customerPaymentInstruments}" var="paymentInstrument" status="loopState">
-    <div class="row saved-payment-instrument" data-uuid="${paymentInstrument.UUID}">
-        <div class="form-group required saved-security-code col-6 col-md-3">
-            <img class="card-image ${loopState.first ? 'checkout-hidden' : ''}"
-                src="${paymentInstrument.cardTypeImage.src}"
-                alt="${paymentInstrument.cardTypeImage.alt}"
-            >
-            <div class="security-code-input ${loopState.first ? '' : 'checkout-hidden'}">
-                <label class="form-control-label" for="saved-payment-security-code">${Resource.msg('label.credit.card-security.code','checkout',null)}</label>
+    <isscript>
+        // Get the saved card
+        var card = savedCardHelper.getSavedCard(
+            paymentInstrument.UUID,
+            BasketMgr.currentBasket.getCustomerNo(),
+            paymentOption.ID
+        );
+    </isscript>
+    <isif condition="${card && card.getCreditCardToken()}">
+        <div class="row saved-payment-instrument" data-uuid="${paymentInstrument.UUID}">
+            <div class="form-group required saved-security-code col-6 col-md-3">
+                <img class="card-image ${loopState.first ? 'checkout-hidden' : ''}"
+                    src="${paymentInstrument.cardTypeImage.src}"
+                    alt="${paymentInstrument.cardTypeImage.alt}"
+                >
+                <div class="security-code-input ${loopState.first ? '' : 'checkout-hidden'}">
+                    <label class="form-control-label" for="saved-payment-security-code">${Resource.msg('label.credit.card-security.code','checkout',null)}</label>
 
-                <button type="button" class="info-icon">
-                    <span class="icon" aria-hidden="true">i</span>
-                    <span class="sr-only">${Resource.msg('button.help.assistive.text','common',null)}</span>
-                    <span class="tooltip d-none">
-                        ${Resource.msg('tooltip.security.code','creditCard',null)}
+                    <button type="button" class="info-icon">
+                        <span class="icon" aria-hidden="true">i</span>
+                        <span class="sr-only">${Resource.msg('button.help.assistive.text','common',null)}</span>
+                        <span class="tooltip d-none">
+                            ${Resource.msg('tooltip.security.code','creditCard',null)}
+                        </span>
+                    </button>
+
+                    <input type="text" class="form-control saved-payment-security-code" id="saved-payment-security-code" maxlength="4"
+                        aria-describedby="savedPaymentSecurityCodeInvalidMessage">
+                    <div class="invalid-feedback" id="savedPaymentSecurityCodeInvalidMessage">
+                    ${Resource.msg('cko.cvv.invalid', 'cko', null)}
+                    </div>
+                </div>
+            </div>
+            <div class="saved-payment-information col-6 col-md-9">
+                <div class="saved-credit-card-type">
+                    <span>
+                        ${Resource.msg('msg.payment.type.credit', 'confirmation', null)}
+                        ${paymentInstrument.creditCardType}
                     </span>
-                </button>
-
-                <input type="text" class="form-control saved-payment-security-code" id="saved-payment-security-code" maxlength="4"
-                    aria-describedby="savedPaymentSecurityCodeInvalidMessage">
-                <div class="invalid-feedback" id="savedPaymentSecurityCodeInvalidMessage">
-                ${Resource.msg('cko.cvv.invalid', 'cko', null)}
+                </div>
+                <div class="saved-credit-card-number">
+                    ${paymentInstrument.maskedCreditCardNumber}
+                </div>
+                <div class="saved-credit-card-expiration-date">
+                    <span>
+                        ${Resource.msg('msg.card.type.ending', 'confirmation', null)}
+                        ${paymentInstrument.creditCardExpirationMonth}/${paymentInstrument.creditCardExpirationYear}
+                    </span>
                 </div>
             </div>
         </div>
-        <div class="saved-payment-information col-6 col-md-9">
-            <div class="saved-credit-card-type">
-                <span>
-                    ${Resource.msg('msg.payment.type.credit', 'confirmation', null)}
-                    ${paymentInstrument.creditCardType}
-                </span>
-            </div>
-            <div class="saved-credit-card-number">
-                ${paymentInstrument.maskedCreditCardNumber}
-            </div>
-            <div class="saved-credit-card-expiration-date">
-                <span>
-                    ${Resource.msg('msg.card.type.ending', 'confirmation', null)}
-                    ${paymentInstrument.creditCardExpirationMonth}/${paymentInstrument.creditCardExpirationYear}
-                </span>
-            </div>
-        </div>
-    </div>
+    </isif>
 </isloop>
