+++ Cartridges/int_checkoutcom_sfra/cartridge/templates/resources/cko.properties
@@ -95,3 +95,4 @@ cko.sepa.paynow=Pay Now
 cko.sepa.cancel=Cancel
 cko.sepa.iban=IBAN
 cko.sepa.bic=BIC
+cko.sepa.tandc=I accept the mandate for a single payment
